refactor: remove textfield and inputsegments from DTP inheritance chain
