+++ packages/calendar/src/index.ts
@@ -10,3 +10,4 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 export * from './Calendar.js';
+export * from './types.js';
