+++ packages/date-time-picker/package.json
@@ -34,6 +34,10 @@
             "development": "./src/index.dev.js",
             "default": "./src/index.js"
         },
+        "./src/types.js": {
+            "development": "./src/types.dev.js",
+            "default": "./src/types.js"
+        },
         "./sp-date-time-picker.js": {
             "development": "./sp-date-time-picker.dev.js",
             "default": "./sp-date-time-picker.js"
@@ -58,15 +62,15 @@
     ],
     "dependencies": {
         "@spectrum-web-components/calendar": "^0.0.1",
-        "@spectrum-web-components/field-label": "^0.41.2",
+        "@spectrum-web-components/help-text": "^0.41.2",
         "@spectrum-web-components/icons-workflow": "^0.41.2",
-        "@spectrum-web-components/input-segments": "^0.0.1",
         "@spectrum-web-components/overlay": "^0.41.2",
         "@spectrum-web-components/picker-button": "^0.41.2",
-        "@spectrum-web-components/popover": "^0.41.2"
+        "@spectrum-web-components/popover": "^0.41.2",
+        "@spectrum-web-components/shared": "^0.41.2"
     },
     "devDependencies": {
-        "@spectrum-web-components/story-decorator": "^0.41.2"
+        "@spectrum-css/datepicker": "^3.2.1"
     },
     "types": "./src/index.d.ts",
     "customElements": "custom-elements.json",
