+++ packages/date-time-picker/src/date-time-picker.css
@@ -9,6 +9,10 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
+
+@import url('./spectrum-date-time-picker.css');
+
+/*
 :host {
     --status-icon-size: var(
         --mod-textfield-icon-size-invalid,
@@ -59,46 +63,109 @@ governing permissions and limitations under the License.
     gap: calc(var(--input-spacing) - var(--input-border));
 }
 
-.picker-button[invalid] {
+sp-picker-button[invalid] {
     --mod-picker-button-border-color: var(
         --spectrum-textfield-border-color-invalid-default
     );
 }
 
-.picker-button[invalid]:focus {
+sp-picker-button[invalid]:focus {
     --mod-picker-button-border-color: var(
         --spectrum-textfield-border-color-invalid-focus
     );
 }
 
-.picker-button[invalid]:focus-visible {
+sp-picker-button[invalid]:focus-visible {
     --mod-picker-button-border-color: var(
         --spectrum-textfield-border-color-invalid-keyboard-focus
     );
 }
 
-.picker-button[invalid]:hover {
+sp-picker-button[invalid]:hover {
     --mod-picker-button-border-color: var(
         --spectrum-textfield-border-color-hover
     );
 }
 
-.picker-button[invalid]:focus:hover {
+sp-picker-button[invalid]:focus:hover {
     --mod-picker-button-border-color: var(
         --spectrum-textfield-border-color-invalid-focus-hover
     );
 }
 
-:host([invalid]:hover) .picker-button[invalid] {
+:host([invalid]:hover) sp-picker-button[invalid] {
     --mod-picker-button-border-color: var(
         --spectrum-textfield-border-color-hover
     );
 }
+*/
 
-.popover {
+sp-popover {
     overflow-y: auto;
 }
 
-.popover-content {
+sp-popover .popover-content {
     padding: var(--spectrum-spacing-300);
+}
+
+.input-content {
+    display: flex;
+    align-items: center;
+    height: 100%;
+    overflow-x: auto;
+    scrollbar-width: none; /* Firefox */
+    line-height: normal;
+}
+
+.input-content::-webkit-scrollbar {
+    display: none;
+    width: 0;
+    height: 0;
+}
+
+.literal-segment,
+.editable-segment {
+    display: inline-block;
+    height: 100%;
+    color: var(--spectrum-textfield-text-color-default);
+}
+
+.editable-segment {
+    padding: 0 0.15em;
+    font-variant-numeric: tabular-nums;
+    text-align: end;
+    white-space: nowrap;
+    caret-color: transparent;
+    outline: none;
+}
+
+[dir='ltr'] .editable-segment {
+    text-align: start;
+}
+
+.literal-segment {
+    white-space: pre;
+    user-select: none;
+}
+
+.is-placeholder,
+.is-placeholder + .literal-segment {
+    color: var(--spectrum-gray-500);
+}
+
+.is-placeholder:not(:is(:lang(ja), :lang(ko), :lang(zh))) {
+    font-style: italic;
+}
+
+.editable-segment:focus {
+    color: var(--spectrum-white);
+    background-color: var(--spectrum-accent-background-color-default);
+}
+
+/**
+ * Hide selection because there is no way to avoid it entirely in Firefox
+ * https://bugzilla.mozilla.org/show_bug.cgi?id=1742153
+ */
+.editable-segment::selection {
+    background-color: transparent;
 }
