+++ packages/date-time-picker/src/index.ts
@@ -10,3 +10,4 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 export * from './DateTimePicker.js';
+export * from './types.js';
