+++ packages/date-time-picker/src/types.ts
@@ -8,6 +8,9 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
+import { DateValue } from '@spectrum-web-components/calendar';
+
+export type DateTimePickerValue = DateValue;
 
 export const dateSegmentTypes: ReadonlyArray<Intl.DateTimeFormatPartTypes> = [
     'day',
