+++ packages/date-time-picker/tsconfig.json
@@ -5,6 +5,5 @@
         "rootDir": "./"
     },
     "include": ["*.ts", "src/*.ts"],
-    "exclude": ["test/*.ts", "stories/*.ts"],
-    "references": [{ "path": "../../tools/input-segments" }]
+    "exclude": ["test/*.ts", "stories/*.ts"]
 }
