--- tools/input-segments/.npmignore
@@ -1,2 +1,0 @@
-stories
-test
