--- tools/input-segments/exports.json
@@ -1,3 +1,0 @@
-{
-    "./src/*": "./src/*.js"
-}
