--- tools/input-segments/package.json
@@ -1,71 +1,0 @@
-{
-    "name": "@spectrum-web-components/input-segments",
-    "version": "0.0.1",
-    "publishConfig": {
-        "access": "public"
-    },
-    "description": "“Abstract” component used to extend date and time form fields classes",
-    "license": "Apache-2.0",
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/adobe/spectrum-web-components.git",
-        "directory": "tools/input-segments"
-    },
-    "author": "",
-    "homepage": "https://adobe.github.io/spectrum-web-components/tools/input-segments",
-    "bugs": {
-        "url": "https://github.com/adobe/spectrum-web-components/issues"
-    },
-    "main": "./src/index.js",
-    "module": "./src/index.js",
-    "type": "module",
-    "exports": {
-        ".": {
-            "development": "./src/index.dev.js",
-            "default": "./src/index.js"
-        },
-        "./package.json": "./package.json",
-        "./src/InputSegments.js": {
-            "development": "./src/InputSegments.dev.js",
-            "default": "./src/InputSegments.js"
-        },
-        "./src/index.js": {
-            "development": "./src/index.dev.js",
-            "default": "./src/index.js"
-        },
-        "./src/input-segments.css.js": "./src/input-segments.css.js",
-        "./src/types.js": {
-            "development": "./src/types.dev.js",
-            "default": "./src/types.js"
-        }
-    },
-    "scripts": {
-        "test": "echo \"Error: run tests from mono-repo root.\" && exit 1"
-    },
-    "files": [
-        "**/*.d.ts",
-        "**/*.js",
-        "**/*.js.map",
-        "custom-elements.json",
-        "!stories/",
-        "!test/"
-    ],
-    "keywords": [
-        "spectrum css",
-        "web components",
-        "lit-element",
-        "lit-html"
-    ],
-    "dependencies": {
-        "@internationalized/date": "^3.5.1",
-        "@internationalized/number": "^3.1.0",
-        "@spectrum-web-components/base": "^0.41.2",
-        "@spectrum-web-components/reactive-controllers": "^0.41.2",
-        "@spectrum-web-components/textfield": "^0.41.2"
-    },
-    "types": "./src/index.d.ts",
-    "customElements": "custom-elements.json",
-    "sideEffects": [
-        "./**/*.dev.js"
-    ]
-}
