--- tools/input-segments/tsconfig.json
@@ -1,14 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-        "composite": true,
-        "rootDir": "./"
-    },
-    "include": ["*.ts", "src/*.ts"],
-    "exclude": ["test/*.ts", "stories/*.ts"],
-    "references": [
-        { "path": "../base" },
-        { "path": "../reactive-controllers" },
-        { "path": "../../packages/textfield" }
-    ]
-}
