+++ tsconfig-all.json
@@ -89,7 +89,6 @@
         { "path": "tools/base" },
         { "path": "tools/bundle" },
         { "path": "tools/grid" },
-        { "path": "tools/input-segments" },
         { "path": "tools/reactive-controllers" },
         { "path": "tools/shared" },
         { "path": "tools/styles" },
